requirements.txt
streamlit==1.28.1
websocket-client==1.6.3
pandas==2.1.1